+++ ChangeLog.md
@@ -1,2 +1,7 @@
   Version 1.0.0
   - First public release
+
+  Version 1.1.0
+  - minor "Cache-Control" parameters order changes
+  - added bpoDelegateIndexToInheritedDefault to delegate index.html to Default()
+  - added bpoDelegate404ToInherited_404 to delegate 404.html to _404()